refactor: split `api.ts`
