+++ src/api/account.ts
@@ -1,0 +1,73 @@
+import { JsonStruct } from '@metamask/utils';
+import type { Infer } from 'superstruct';
+import { array, enums, record, string } from 'superstruct';
+
+import { object } from '../superstruct';
+import { UuidStruct } from '../utils';
+
+/**
+ * Supported Ethereum account types.
+ */
+export enum EthAccountType {
+  Eoa = 'eip155:eoa',
+  Erc4337 = 'eip155:erc4337',
+}
+
+/**
+ * Supported Bitcoin account types.
+ */
+export enum BtcAccountType {
+  P2wpkh = 'bip122:p2wpkh',
+}
+
+/**
+ * Supported account types.
+ */
+export type KeyringAccountType =
+  | `${EthAccountType.Eoa}`
+  | `${EthAccountType.Erc4337}`
+  | `${BtcAccountType.P2wpkh}`;
+
+/**
+ * A struct which represents a Keyring account object. It is abstract enough to
+ * be used with any blockchain. Specific blockchain account types should extend
+ * this struct.
+ *
+ * See {@link KeyringAccount}.
+ */
+export const KeyringAccountStruct = object({
+  /**
+   * Account ID (UUIDv4).
+   */
+  id: UuidStruct,
+
+  /**
+   * Account type.
+   */
+  type: enums([
+    `${EthAccountType.Eoa}`,
+    `${EthAccountType.Erc4337}`,
+    `${BtcAccountType.P2wpkh}`,
+  ]),
+
+  /**
+   * Account main address.
+   */
+  address: string(),
+
+  /**
+   * Account options.
+   */
+  options: record(string(), JsonStruct),
+
+  /**
+   * Account supported methods.
+   */
+  methods: array(string()),
+});
+
+/**
+ * Keyring Account type represents an account and its properties from the
+ * point of view of the keyring.
+ */
+export type KeyringAccount = Infer<typeof KeyringAccountStruct>;
