+++ src/api/keyring.ts
@@ -1,0 +1,142 @@
+import type { Json } from '@metamask/utils';
+
+import type { KeyringAccount } from './account';
+import type { KeyringAccountData } from './export';
+import type { KeyringRequest } from './request';
+import type { KeyringResponse } from './response';
+
+/**
+ * Keyring interface.
+ *
+ * Represents the functionality and operations related to managing accounts and
+ * handling requests.
+ */
+export type Keyring = {
+  /**
+   * List accounts.
+   *
+   * Retrieves an array of KeyringAccount objects representing the available
+   * accounts.
+   *
+   * @returns A promise that resolves to an array of KeyringAccount objects.
+   */
+  listAccounts(): Promise<KeyringAccount[]>;
+
+  /**
+   * Get an account.
+   *
+   * Retrieves the KeyringAccount object for the given account ID.
+   *
+   * @param id - The ID of the account to retrieve.
+   * @returns A promise that resolves to the KeyringAccount object if found, or
+   * undefined otherwise.
+   */
+  getAccount(id: string): Promise<KeyringAccount | undefined>;
+
+  /**
+   * Create an account.
+   *
+   * Creates a new account with optional, keyring-defined, account options.
+   *
+   * @param options - Keyring-defined options for the account (optional).
+   * @returns A promise that resolves to the newly created KeyringAccount
+   * object without any private information.
+   */
+  createAccount(options?: Record<string, Json>): Promise<KeyringAccount>;
+
+  /**
+   * Filter supported chains for a given account.
+   *
+   * @param id - ID of the account to be checked.
+   * @param chains - List of chains (CAIP-2) to be checked.
+   * @returns A Promise that resolves to a filtered list of CAIP-2 IDs
+   * representing the supported chains.
+   */
+  filterAccountChains(id: string, chains: string[]): Promise<string[]>;
+
+  /**
+   * Update an account.
+   *
+   * Updates the account with the given account object. Does nothing if the
+   * account does not exist.
+   *
+   * @param account - The updated account object.
+   * @returns A promise that resolves when the account is successfully updated.
+   */
+  updateAccount(account: KeyringAccount): Promise<void>;
+
+  /**
+   * Delete an account from the keyring.
+   *
+   * Deletes the account with the given ID from the keyring.
+   *
+   * @param id - The ID of the account to delete.
+   * @returns A promise that resolves when the account is successfully deleted.
+   */
+  deleteAccount(id: string): Promise<void>;
+
+  /**
+   * Exports an account's private key.
+   *
+   * If the keyring cannot export a private key, this function should throw an
+   * error.
+   *
+   * @param id - The ID of the account to export.
+   * @returns A promise that resolves to the exported account.
+   */
+  exportAccount?(id: string): Promise<KeyringAccountData>;
+
+  /**
+   * List all submitted requests.
+   *
+   * Retrieves an array of KeyringRequest objects representing the submitted
+   * requests.
+   *
+   * @returns A promise that resolves to an array of KeyringRequest objects.
+   */
+  listRequests?(): Promise<KeyringRequest[]>;
+
+  /**
+   * Get a request.
+   *
+   * Retrieves the KeyringRequest object for the given request ID.
+   *
+   * @param id - The ID of the request to retrieve.
+   * @returns A promise that resolves to the KeyringRequest object if found, or
+   * undefined otherwise.
+   */
+  getRequest?(id: string): Promise<KeyringRequest | undefined>;
+
+  /**
+   * Submit a request.
+   *
+   * Submits the given KeyringRequest object.
+   *
+   * @param request - The KeyringRequest object to submit.
+   * @returns A promise that resolves to the request response.
+   */
+  submitRequest(request: KeyringRequest): Promise<KeyringResponse>;
+
+  /**
+   * Approve a request.
+   *
+   * Approves the request with the given ID and sets the response if provided.
+   *
+   * @param id - The ID of the request to approve.
+   * @param data - The response to the request (optional).
+   * @returns A promise that resolves when the request is successfully
+   * approved.
+   */
+  approveRequest?(id: string, data?: Record<string, Json>): Promise<void>;
+
+  /**
+   * Reject a request.
+   *
+   * Rejects the request with the given ID.
+   *
+   * @param id - The ID of the request to reject.
+   * @returns A promise that resolves when the request is successfully
+   * rejected.
+   */
+  rejectRequest?(id: string): Promise<void>;
+};
